+++ package.json
@@ -10,7 +10,7 @@
     "build": "tsc --project ./tsconfig.json --module commonjs --outDir ./src/dist --declaration",
     "dev": "yarn build --watch",
     "generate": "wagmi generate",
-    "generate:node-config-type": "node ./src/dist/scripts/generateNodeConfigType.js && prettier --write ./src/types/NodeConfig-Generated.ts",
+    "generate:node-config-type": "node ./src/dist/scripts/generateNodeConfigType.js && prettier --write ./src/types/NodeConfig.generated.ts",
     "test:unit": "vitest unit.test",
     "test:integration": "vitest integration.test",
     "postinstall": "patch-package",
