+++ src/types/NodeConfig.generated.ts
@@ -3,7 +3,7 @@
 // THIS FILE IS AUTOMATICALLY GENERATED AND SHOULD NOT BE EDITED MANUALLY
 //
 // IMAGE: offchainlabs/nitro-node:v2.1.2-4c55843-dev
-// TIMESTAMP: 2024-01-05T15:52:26.764Z
+// TIMESTAMP: 2024-01-05T16:01:47.230Z
 //
 // ---
 /** Nitro node configuration options */
